+++ aplicacion/telefonia/__init__.py
@@ -1,0 +1,1 @@
+"""Control de telefonía Asterisk mediante ARI."""